backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
pydantic==2.11.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.31.0
alembic==1.13.1
Werkzeug==3.0.1

# Additional deployment / runtime dependencies
# Mistral AI client used by `ai_service.py`
mistralai==0.4.2

# WSGI servers for deployment (choose one for your target platform):
# - Use `gunicorn` on Linux/Unix.
# - Use `waitress` on Windows or simple deployments.
gunicorn==20.1.0
waitress==2.1.2

# Ensure SQLAlchemy core matches Flask-SQLAlchemy expectations
SQLAlchemy==2.0.20

# Backports / typing helpers required by some libraries
# typing_extensions==4.8.0

# Optional: helpful utilities for URL validation and env management
validators==0.20.0
